+++ ContaCertaApi/Configs/ServicesExtensions.cs
@@ -1,46 +1,14 @@
-﻿using ContaCerta.Api.Infra.Context;
-using ContaCerta.Api.Infra.Costs;
-using ContaCerta.Api.Infra.Users;
-using ContaCerta.Domain.Costs.Repositories.Interfaces;
-using ContaCerta.Domain.Costs.Services;
-using ContaCerta.Domain.Costs.Validates;
-using ContaCerta.Domain.Costs.Validates.Interfaces;
-using ContaCerta.Domain.Users.Repositories.Interfaces;
-using ContaCerta.Domain.Users.Services;
-using ContaCerta.Domain.Users.Validates;
-using ContaCerta.Domain.Users.Validates.Interfaces;
+﻿using ContaCerta.Infra.Context;
 using Microsoft.EntityFrameworkCore;
 
-namespace ContaCerta.Api.Configs
+namespace ContaCerta.Api.Configs;
+
+public static class ServicesExtensions
 {
-    public static class ServicesExtensions
+    public static IServiceCollection RegisterContexts(this IServiceCollection Services, string? stringConnection)
     {
-        public static IServiceCollection RegisterContexts(this IServiceCollection Services, string? stringConnection)
-        {
-            Services.AddDbContext<ContaCertaContext>(options =>
-                options.UseSqlServer(stringConnection));
-            return Services;
-        }
-
-        public static IServiceCollection RegisterServices(this IServiceCollection Services)
-        {
-            Services.AddTransient<ManagerCost>();
-            Services.AddTransient<ManagerUsersInCost>();
-            Services.AddTransient<ManagerUser>();
-            Services.AddTransient<ListCostsUser>();
-
-            return Services;
-        }
-
-        public static IServiceCollection ResolveInterfaces(this IServiceCollection Services)
-        {
-            Services.AddTransient<ICostRepository, CostRepository>();
-            Services.AddTransient<IUserRepository, UserRepository>();
-            Services.AddTransient<ICostValidate, CostValidate>();
-            Services.AddTransient<IEmailValidate, EmailValidate>();
-            Services.AddTransient<IPasswordValidate, PasswordValidate>();
-            Services.AddTransient<IUserCostRepository, UserCostRepository>();
-            return Services;
-        }
+        Services.AddDbContext<ContaCertaContext>(options =>
+            options.UseSqlServer(stringConnection));
+        return Services;
     }
 }
